playify/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.4
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.4
djangorestframework==3.12.2
entrypoints==0.3
filelock==3.0.12
idna==2.9
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
isort==5.4.2
jedi==0.17.1
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyter-http-over-ws==0.0.8
jupyterlab==2.1.4
jupyterlab-server==1.1.5
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
pycodestyle==2.6.0
Pygments==2.6.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.4
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.1
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
virtualenv==20.0.31
wcwidth==0.2.4
webencodings==0.5.1
wrapt==1.12.1
